make rating double, change weights from 0.1 to 1.0

+++ ticket-backend/src/main/java/aics/domain/fuzzy/models/FuzzyVariableDistributionPart.java
@@ -18,9 +18,9 @@ import java.io.Serializable;
 public abstract class FuzzyVariableDistributionPart implements Serializable {
     FuzzyVariableDistributionType type;
 
-    abstract public int findFirstValue();
+    abstract public double findFirstValue();
 
-    abstract public int findLastValue();
+    abstract public double findLastValue();
 
     abstract public double calculateFuzzyPartMx(double x);
