+++ ticket-backend/src/main/java/aics/domain/fuzzy/models/FuzzyVariableDistributionPartTrapezoidal.java
@@ -17,12 +17,12 @@ import java.util.Objects;
 @Accessors(chain = true)
 public class FuzzyVariableDistributionPartTrapezoidal extends FuzzyVariableDistributionPart implements Serializable {
     private String partName;
-    private Integer a;
-    private int b;
-    private int c;
-    private Integer d;
+    private Double a;
+    private double b;
+    private double c;
+    private Double d;
 
-    public FuzzyVariableDistributionPartTrapezoidal(String partName, Integer a, int b, int c, Integer d) {
+    public FuzzyVariableDistributionPartTrapezoidal(String partName, Double a, double b, double c, Double d) {
         this.setType(FuzzyVariableDistributionType.TRAPEZOIDAL);
         this.partName = partName;
         this.a = a;
@@ -33,12 +33,12 @@ public class FuzzyVariableDistributionPartTrapezoidal extends FuzzyVariableDistr
 
 
     @Override
-    public int findFirstValue() {
+    public double findFirstValue() {
         return Objects.requireNonNullElse(this.a, this.b);
     }
 
     @Override
-    public int findLastValue() {
+    public double findLastValue() {
         return Objects.requireNonNullElse(this.d, this.c);
     }
